Correctness fix for RelativeDistinguishedName::eq
Sort the NameTypeValue objects by type and stringprep'd value, rather
than by DER bytes. If a RDN had two sequences with the same type, and
different values, and the values were encoded differently in different
certificates, sorting the RDN set by DER bytes would produce a different
ordering depending on the encoding, and the two encodings of the same
RDN would incorrectly be treated as not equal.