Fix chain building with many self-signed certs
If there were many different self-signed certificates for the same name,
chain building would take far too long, as it was exploring paths with
the self-signed certificates in every possible order. Instead,
self-signed certificates won't be considered during chain building.
(With the obvious exception of trust roots that are self-signed)